Rename cli package to tgpu-gen (#505)
