+++ packages/tgpu-gen/colors.mjs
