+++ packages/tgpu-gen/outputPathCompiler.mjs
