+++ packages/tgpu-gen/tests/structs.test.ts
